Visibly mark impossible exits
For example, if a Boss entrance is assigned to an exit, but the Dungeon
exit for that Boss has not been assigned, then the exit at that Boss
entrance is impossible. This may happen if the user has decided to not
randomize Boss/Miniboss entrances, but has randomized Dungeon entrances
because the user can fill in the Boss/Miniboss entrances from the start.