Add swift sail sequence break to can_fly_with_deku_leaf_outdoors
There are a few cases where the wind must be changed without access to
KoRL, so these cannot be done with this sequence break:

To get into Forest Haven without the Grappling Hook, the wind must be
a specific direction. To do the checks without Grappling hook or to
access the dungeon entrance requires flying out to the pillar between
Forest Haven and Forbidden Woods, which requires a different wind
direction, but Swift Sail can only set the wind direction from the boat.

Flying around the islands around Dragon Roost Island requires changing
the wind direction multiple times, so this is not possible either.

There may be more that have been missed, but these will be fixed when
they are found.