Fix processing map changes for other slots
The `slots` and `tags` in Bounce messages actually specify to send to
all clients that match either the slot or the tag, so the client is
currently mistakenly sending to all connected trackers in addition to
all clients connected to the player's slot.

A client fix to remove sending Bounced messages to all clients with the
"Tracker" tag can come later, but checking the Bounced message contains
our slot fixes the issue on the tracker's side.